requirements.txt
beautifulsoup4==4.14.2
fake_useragent==2.2.0
pandas==2.3.3
protobuf==5.29.5
python-dotenv==1.2.1
selenium==4.38.0
streamlit==1.51.0
webdriver_manager==4.0.2
